django-api/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
sentence-transformers==2.2.2
uwsgi>=2.0.19<2.1